Move query to new separate ix_apps package
